requirements.txt
python-telegram-bot==20.0
eth-account==0.6.0
web3==6.0.0
mnemonic==0.20